# src/main/resources/swagger.yaml
swagger: '2.0'
info:
  title: Event Planning App
  version: 1.0.0
tags:
  - name: Events
    description: API for managing events
  - name: Invites
    description: API for managing invites
  - name: Login
    description: API for Login
  - name: Signup
    description: API for Signup

definitions:
  Event:
    type: object
    properties:
      id: 
        type: integer
        format: int64
      nome:
        type: string
        description: Name of the event
      date:
        type: string
        format: date-time
        description: Date of the event
      place:
        type: string
        description: Location of the event
      guests:
        type: array
        description: List of guests
        items:
          type: string
          
  Invite:
    type: object
    properties:
      id:
        type: integer
        format: int64
      name:
        type: string
        description: Name of the invited guest
      email:
        type: string
        description: Email address of the invited guest
        
  LoginRequest:
    type: object
    required:
      - "username"
      - "password"
    properties:
      id:
        type: integer
      username: 
        type: string
      password:
        type: string
      phone:
        type: string
      email:
        type: string
      roles:
        type: array
        items:
          type: string

  SignupRequest:
    type: object
    required:
      - "username"
      - "password"
    properties:
      id:
        type: integer
      username: 
        type: string
      password:
        type: string
      phone:
        type: string
      email:
        type: string
      roles:
        type: array
        items:
          type: string
        
  ErrorModel:
    type: object
    required:
      - "errorCode"
      - "errorMessage"
    properties:
      errorCode: 
        type: string
      errorMessage:
        type: string
      errors:
        type: object
        additionalProperties:
          type: object

  UserInfoResponse:
    type: object
    properties:
      id: 
        type: integer
        format: int64
      username:
        type: string
      email: 
        type: string
      phone: 
        type: string
      roles:
        type: array
        items:
          type: string
        
      
paths:
  /event:
    post:
      summary: Create a new event
      operationId: createEvent
      tags:
        - Events
      parameters:
        - in: body
          name: body
          description: Data of the event to be created
          required: true
          schema:
            $ref: '#/definitions/Event'
      responses:
        201:
          description: The event has been created successfully
    get:
      summary: Retrieve all events
      tags:
        - Events
      responses:
        200:
          description: List of all events
          schema:
            type: array
            items:
              $ref: '#/definitions/Event'
              
  /event/{eventId}:
    get:
      tags:
        - Events
      summary: Retrieve a specific event
      operationId: retrieveEvents
      parameters:
        - in: path
          name: eventId
          required: true
          type: string
          description: ID of the event to retrieve
      responses:
        200:
          description: Details of the event
          schema:
            $ref: '#/definitions/Event'
    put:
      tags:
        - Events
      summary: Update a specific event
      operationId: updateEvent
      parameters:
        - in: path
          name: eventId
          required: true
          type: string
          description: ID of the event to update
        - in: body
          name: body
          description: Data of the event to update
          required: true
          schema:
            $ref: '#/definitions/Event'
      responses:
        200:
          description: The event has been updated successfully
    delete:
      tags:
        - Events
      summary: Delete a specific event
      operationId: deleteEvent
      parameters:
        - in: path
          name: eventId
          required: true
          type: string
          description: ID of the event to delete
      responses:
        204:
          description: The event has been deleted successfully
  
  /event/{eventId}/invite:
    post:
      tags:
        - Invites
      summary: Invite a guest to the event
      operationId: createInvite
      parameters:
        - in: path
          name: eventId
          required: true
          type: string
          description: ID of the event to send the invitation to
        - in: body
          name: body
          description: Data of the invitation to send
          required: true
          schema:
            $ref: '#/definitions/Invite'
      responses:
        201:
          description: The invitation has been sent successfully
  /event/{eventId}/invite/{inviteId}:
    get:
      tags:
        - Invites
      summary: Retrieve a specific invitation
      operationId: retrieveSpecificInvite
      parameters:
        - in: path
          name: eventId
          required: true
          type: string
        - in: path
          name: inviteId
          required: true
          type: string
      responses:
          204:
            description: L'invito è stato cancellato con successo
            
  /login:
    post:
      tags:
        - Login
      summary: Login API
      operationId: login
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/LoginRequest'
      responses:
        200:
          description: Success
          schema:
            $ref: '#/definitions/UserInfoResponse'
        500:
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorModel'
        401:
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorModel'
        400:
          description: BadRequest
          schema:
            $ref: '#/definitions/ErrorModel'
            
  /signup:
    post:
      tags:
        - Signup
      summary: Signup API
      operationId: signup
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/SignupRequest'
      responses:
        200:
          description: Success
          schema:
            $ref: '#/definitions/UserInfoResponse'
        500:
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorModel'
        401:
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorModel'
        400:
          description: BadRequest
          schema:
            $ref: '#/definitions/ErrorModel'